dwm - Apply patch for window following
Provides "window following" behaviour, which causes the tag/monitor
a window is moved to with tag or tagmon to be focused. The behaviour is
affected in strange ways by which monitor the cursor is on, but this is
true for dwm normally.

An icon indicating whether this behaviour is active is added to the bar:

v means window following is off.
> means window following is on.
The togglefollow function toggles whether this occurs on the current monitor.
By default, this is triggered by MODKEY+n and by clicking the icon.

The WFDEFAULT macro in config.def.h is used to determine whether window
following should be on by default, and the values of WFACTIVE and WFINACTIVE
can be changed to other char values (which must be different) so the icon
looks different.

dwm-windowfollow-20221002-69d5652.diff (Working on 6.4)